last update

+++ carts/views.py
@@ -11,14 +11,12 @@ def _cart_id(request):
     if not cart:
         cart = request.session.create()
     return cart
-
-
 def add_cart(request, product_id):
-    product = Product.objects.get(id=product_id)
+    product = get_object_or_404(Product, id=product_id)
     product_variation = []
 
     if request.method == "POST":
-        for key, value in request.POST.items():  
+        for key, value in request.POST.items():
             try:
                 variation = Variation.objects.get(
                     product=product,
@@ -29,49 +27,58 @@ def add_cart(request, product_id):
             except Variation.DoesNotExist:
                 pass
 
-    try:
-        cart = Cart.objects.get(cart_id=_cart_id(request))
-    except Cart.DoesNotExist:
-        cart = Cart.objects.create(cart_id=_cart_id(request))
+    cart, created = Cart.objects.get_or_create(cart_id=_cart_id(request))
 
-    try:
-        cart_item = CartItem.objects.get(product=product, cart=cart)
+    is_cart_item_exists = CartItem.objects.filter(product=product, cart=cart, variations__in=product_variation).exists()
+
+    if is_cart_item_exists:
+        cart_item = CartItem.objects.filter(product=product, cart=cart, variations__in=product_variation).first()
         cart_item.quantity += 1
         cart_item.save()
-    except CartItem.DoesNotExist:
+    else:
         cart_item = CartItem.objects.create(
             product=product,
             quantity=1,
             cart=cart,
         )
+        if len(product_variation) > 0:
+            cart_item.variations.add(*product_variation)
 
     return redirect("cart")
 
 
-def remove_cart(request, product_id):
+def remove_cart(request, product_id, cart_item_id):
+    cart = Cart.objects.get(cart_id=_cart_id(request))
     product = Product.objects.get(id=product_id)
 
     try:
-        cart = Cart.objects.get(cart_id=_cart_id(request))
-        cart_item = CartItem.objects.get(product=product, cart=cart)
+        cart_item = CartItem.objects.filter(product=product, cart=cart, id=cart_item_id).first()
 
-        if cart_item.quantity > 1:
-            cart_item.quantity -= 1
-            cart_item.save()
-        else:
-            cart_item.delete()
+        if cart_item:
+            if cart_item.quantity > 1:
+                cart_item.quantity -= 1
+                cart_item.save()
+            else:
+                cart_item.delete()
 
-    except (Cart.DoesNotExist, CartItem.DoesNotExist):
+    except CartItem.DoesNotExist:
         pass
 
     return redirect("cart")
 
-
-def remove_cart_item(request, product_id):
+def remove_cart_item(request, product_id, cart_item_id):
     cart = Cart.objects.get(cart_id=_cart_id(request))
     product = get_object_or_404(Product, id=product_id)
-    cart_item = CartItem.objects.get(product=product, cart=cart)
-    cart_item.delete()
+
+    try:
+        cart_item = CartItem.objects.filter(product=product, cart=cart, id=cart_item_id).first()
+
+        if cart_item:
+            cart_item.delete()
+
+    except CartItem.DoesNotExist:
+        pass
+
     return redirect("cart")
